Fix Several Calculator Issues
In any physical, real-world calculator I can recall using, if I wanted to compute 2+2 and 3+3 I would click the following buttons:

2 + 2 = 3 + 3 =

But if I do that in this program,

2 + 2 =

shows "4", as expected, but then clicking 3 changes the display to "43".

Also, entering

1 / 0 =

displays "error", as expected.  If I then decide to compute 2 + 2, when I click "2" the display changes to "error2".

I appreciate that this sample is intended as an example of using VisualRuby, not an example of how to write calculator logic.  But these issues make the calculator less usable.

This patch fixes the above two issues.